+++ src/transformer/transformChinese.ts
@@ -1,6 +1,6 @@
 import * as ts from 'typescript'
 import { DOUBLE_BYTE_REGEX } from '../const'
-import { genKey, removeFileComment, saveFile } from '../utils'
+import { genKey, removeFileComment, saveFile, getVariableFromTmeplateString } from '../utils'
 
 export interface Text {
   key: string;
@@ -8,8 +8,22 @@ export interface Text {
   comment: string;
 }
 
+interface IConfig {
+  outputPath: string;
+  rootPath: string;
+  template: boolean;
+  extractOnly: boolean;
+  whiteList: string[];
+  mode?: "sample" | "depth"; // 模式类型 简单模式/深层次导出
+  prefix?: string[];
+  // 用于处理模板字符串的配置
+  templateString?: {
+    funcName: string
+  }
+}
+
 export function transformChinese(code: string, fileName: string) {
-  const { extractOnly, prefix } = global["intlConfig"]
+  const { extractOnly, prefix, templateString } = <IConfig>global["intlConfig"]
   const matches: Array<Text> = [];
   const ast = ts.createSourceFile('', code, ts.ScriptTarget.ES2015, true, ts.ScriptKind.TSX)
   const key = genKey(fileName)
@@ -30,7 +44,7 @@ export function transformChinese(code: string, fileName: string) {
             // 原处修改
             if (!extractOnly) {
               const parentNodeKind = node.parent.kind
-              const result = parentNodeKind === ts.SyntaxKind.JsxAttribute ? `{I18N.${name}}` : `I18N.${name}`
+              const result = parentNodeKind === ts.SyntaxKind.JsxAttribute ? `{i18n.${name}}` : `i18n.${name}`
               return ts.createIdentifier(result)
             }
           }
@@ -47,7 +61,7 @@ export function transformChinese(code: string, fileName: string) {
             })
             index += 1
             if (!extractOnly) {
-              return ts.createJsxText(`{ I18N.${name} }`)
+              return ts.createJsxText(`{ i18n.${name} }`)
             }
           }
           break
@@ -56,12 +70,21 @@ export function transformChinese(code: string, fileName: string) {
           const { pos, end } = node
           const text = code.slice(pos, end)
           if (text.match(DOUBLE_BYTE_REGEX)) {
-            console.warn(`模板字符串：${fileName} ${text} 无法处理`);
-            // matches.push({
-            //   key: name,
-            //   value: text,
-            //   comment: `/** ${text} **/`
-            // })
+            if (templateString && templateString.funcName) {
+              // 记录文本匹配
+              matches.push({
+                key: name,
+                value: text.replace(/\$(?=\{)/g, ''), // 先行断言，去掉`$`
+                comment: `/** ${text} **/`
+              })
+              index += 1
+              // 返回新的节点(函数调用)
+              const variableList: string[] = getVariableFromTmeplateString(text)
+              const objParam = ts.createObjectLiteral(variableList.map(variable => ts.createPropertyAssignment(variable, ts.createIdentifier(variable))))
+              return ts.createCall(ts.createIdentifier(templateString.funcName), undefined, [ts.createIdentifier(`i18n.${name}`), objParam])
+            } else {
+              console.warn(`模板字符串：${fileName} ${text} 无法处理`);
+            }
           }
           break
         }